Restrict to a particular version of javax.annotation as part of the build.
This is a total hack to exclude the javax.annotation from being brought in
from eclipse.  The only true way to get arround it is to entirely exclude
eclipse-orbit from the build but there are many other dependencies that
need to be done.

The other option is to contribute back to eclipse orbit and update their
repository so it contains everything we need.